main program
whoops, wrong file previously. only had some of the fixes specified.

fixed the file output and fixed some more grammar and naming mistakes. titles now specify further on what info to input. made the data not input if the input was invalid, and made a messagebox pop up informing the user that the input is invalid.